Mark non-flattened TPM2B's as parameter encryptable
Some TPM structures are flattened for convenience, but not
TPM2B_PRIVATE. The logic that detects whether a type is sized (and
therefore suitable for parameter encryption) depends on the flattening,
which means it misses TPM2B_PRIVATE as a parameter-encryptable type.

This change introduces a workaround into the .NET generator so that it
marks even types in `DontFlatten` properly for parameter encryption.